SSOR-157 restructure project files
