+++ src/app/app.module.ts
@@ -25,8 +25,9 @@ import { ViewProfileComponent } from './user/profile/view-profile/view-profile.c
 
 import { UpdateProfileComponent } from './user/profile/update-profile/update-profile.component';
 import { HomeComponent } from './home/home.component';
-import { RegisterComponent } from './user/register/register.component';
+import { RegisterComponent } from './login/register/register.component';
 import { LoginComponent } from './login/login.component';
+import { ForgotPasswordComponent } from './login/forgot-password/forgot-password.component';
 
 
 @NgModule({
@@ -46,6 +47,7 @@ import { LoginComponent } from './login/login.component';
     TransactionFilterComponent,
     TransactionQueriesComponent,
     LoginComponent,
+    ForgotPasswordComponent,
   ],
   imports: [
     BrowserModule,
