+++ src/app/login/forgot-password/forgot-password.component.html
@@ -1,0 +1,1 @@
+<p>forgot-password works!</p>
